Pull Request 5
Added new unit tests for StringExtensions intended to verify whether its many string parsing methods function properly. Also added two more tests to LineProcessorTests so as to verify that empty Strings throw exceptions in certain methods.  A noticably large amount of code in this extension is heavy on dependencies so a lot of research will likely be necessary to test more complex classes.